Simplified fps config, standardized to client defining its fps, and server only has fallback. Also set fps to float type overall

+++ src/iris/client/config.py
@@ -21,7 +21,7 @@ class VideoConfig(BaseModel):
         ge=240,
         le=1080,
     )
-    capture_fps: int = Field(
+    capture_fps: float = Field(
         default=_yaml_config.get("client", {}).get("video", {}).get("capture_fps", 10),
         ge=1,
         le=30,
